+++ examples/advanced_examples/holdout_test_datasets_example.py
@@ -1,4 +1,4 @@
-from hyperparameter_hunter import Environment, CrossValidationExperiment
+from hyperparameter_hunter import Environment, CVExperiment
 from hyperparameter_hunter.utils.learning_utils import get_toy_classification_data
 from sklearn.model_selection import StratifiedKFold
 from xgboost import XGBClassifier
@@ -30,7 +30,7 @@ def execute():
         cross_validation_params=dict(n_splits=5, shuffle=True, random_state=32),
     )
 
-    experiment = CrossValidationExperiment(
+    experiment = CVExperiment(
         model_initializer=XGBClassifier, model_init_params=dict(subsample=0.5)
     )
     # At the end of the Experiment, notice a few differences from the results of an Experiment given only training data: