+++ hyperparameter_hunter/__init__.py
@@ -26,7 +26,7 @@ except Exception:
 # Set __all__
 ##################################################
 from .environment import Environment
-from .experiments import CrossValidationExperiment
+from .experiments import CVExperiment, CrossValidationExperiment
 from .optimization import BayesianOptimization
 from .optimization import GradientBoostedRegressionTreeOptimization, GBRT
 from .optimization import RandomForestOptimization, RF
@@ -41,6 +41,7 @@ __all__ = [
     #################### Environment ####################
     "Environment",
     #################### Experimentation ####################
+    "CVExperiment",
     "CrossValidationExperiment",
     #################### Hyperparameter Optimization ####################
     "BayesianOptimization",